requirements.txt
google-api-python-client==2.108.0
python-dotenv==1.0.0
streamlit==1.29.0
pandas==2.1.4
isodate==0.6.1
apscheduler==3.10.4
filelock==3.13.1
pytest==7.4.3
